+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # [Versions](https://github.com/Tracktor/design-system-tracktor/releases)
 
 ## v1.11.0
-- Add button story
-- Update dropdown menu style
+- Update paper radius
+- Add new paper stories